change courtage from int to float

+++ modules/70/1/output.php
@@ -890,7 +890,7 @@ if (filter_input(INPUT_GET, 'property_id', FILTER_VALIDATE_INT, ['options' => ['
         echo '<div class="col-12">';
         $real_estate_tax = $d2u_immo->getConfig('finance_calculator_real_estate_tax');
         $notary_costs = $d2u_immo->getConfig('finance_calculator_notary_costs');
-        $courtage = is_numeric($property->courtage) ? ((int) strtr((string) ((int) strtr($property->courtage, ',', '.') * 1), ',', '.') / 100) : 0;
+        $courtage = is_numeric($property->courtage) ? ((float) strtr((string) ((float) strtr($property->courtage, ',', '.') * 1), ',', '.') / 100) : 0;
         $interest_rate = $d2u_immo->getConfig('finance_calculator_interest_rate');
         $repayment = $d2u_immo->getConfig('finance_calculator_repayment');
